refactor: Updated apollo server plugin to break out relevant functions into helpers, utilities, and plugin hooks to reduce cognitive complexity
