+++ lib/query-utils.js
@@ -7,7 +7,7 @@
 
 const OBFUSCATION_STR = '***'
 
-const cleanQuery = (query, argLocations) => {
+module.exports = function cleanQuery(query, argLocations) {
   let cleanedQuery = query
   let offset = 0
 
@@ -22,5 +22,3 @@ const cleanQuery = (query, argLocations) => {
 
   return cleanedQuery
 }
-
-module.exports = cleanQuery
